gem: update Gemfile.lock
whenever was first introduced in development group, and that updated
the Gemfile.lock. Moving it to the production group didn't update the
Gemfile.lock, since it is agnostic to dependencies groups. This in
turn made Travis not update its cache, and miss the whenever gem on
subsequent builds.

I think.