theme: only time-constraint the Emacs process
Some themes (i.e doom-themes) contains a substantial amount of
variants which means the screenshotting takes a while but then the app
picking up/attaching all the screenshots also adds on top of that and
that gets killed by the timeout guard clause.

Since the timeout's purpose is essentially here to catch the Emacs
process, rescope it to just that.