fix: prevent negative page offsets which crash PostgreSQL
How do people manage to access a `themes?page=0` URL is beyond
me; most of the incidents seem to occur on Opera 52, which was a while
ago (march 2018).

Anyhow, I couldn't reproduce the bug locally but eventually figured
that SQLite (which is used in test & development) actually allows:

SELECT * FROM themes LIMIT 3 OFFSET -10;

whereas PostgreSQL (in prod) will promptly throw an exception if a
negative offset is passed.

Redirect to the home page with a notice when a page < 1 is passed.